requirements.txt
pip
pandas
os
numpy
keybert
tqdm
sentence_transformers
contractions
python-dotenv
swifter